+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [Unreleased]
 
 - Add guide support to armies
+- Add ability to comment on armies
 - Add ability to import units from clash of clans army link
 - Add new "Magic Mirror" archer queen equipment
 - Improve and extend filtering UI/capabilities